+++ tablet_qt/menu/settingsmenu.cpp
@@ -425,6 +425,8 @@ OpenableWidget* SettingsMenu::configureServer(CamcopsApp& app)
     });
     page->setTitle(tr("Configure server settings"));
     page->setType(QuPage::PageType::Config);
+    page->registerValidator(std::bind(&SettingsMenu::validateServerSettings,
+                                      this, std::placeholders::_1));
 
     auto questionnaire = new Questionnaire(m_app, {page});
     questionnaire->setFinishButtonIconToTick();
@@ -433,6 +435,19 @@ OpenableWidget* SettingsMenu::configureServer(CamcopsApp& app)
     connect(questionnaire, &Questionnaire::cancelled,
             &app, &CamcopsApp::clearCachedVars);
     return questionnaire;
+}
+
+
+bool SettingsMenu::validateServerSettings(const QuPage* page)
+{
+    Q_UNUSED(page);
+    // Note that we are using cached server variables.
+    const QString hostname = m_app.getCachedVar(varconst::SERVER_ADDRESS).toString();
+    if (hostname.contains("/")) {
+        uifunc::alert("No forward slashes ('/') permitted in server hostname");
+        return false;
+    }
+    return true;
 }